Refactor to provide mechanism for generating zip files rather than only files that represent a table of data
